+++ src/main/java/com/sb02/blogpractice/controller/PostController.java
@@ -33,25 +33,10 @@ public class PostController {
     private final UserService userService;
 
     @PostMapping
-    public ResponseEntity<PostResponseDTO> createPost(@Valid @RequestBody CreatePostRequestDTO createPostRequestDTO,
-                                                      HttpServletRequest httpRequest) {
+    public ResponseEntity<PostResponseDTO> createPost(@Valid @RequestBody CreatePostRequestDTO createPostRequestDTO, HttpServletRequest httpRequest) {
         validateTags(createPostRequestDTO);
         String userId = (String) httpRequest.getAttribute("userId");
-
-        Post post = postService.create(createPostRequestDTO, userId);
-        PostResponseDTO postResponseDTO = entityToDTO(post);
-
-        // 프론트단에서 Content에 업로드된 이미지를 포함해서 내려준다고 가정?
-        extractImageIdsFromContent(createPostRequestDTO.content())
-                        .forEach(id -> {
-                            // 내용의 이미지와 업로드된 이미지가 동일한지 확인 -> 예외 터지면 PostImage 만들지 않고 그냥 포스팅만 진행
-                            try {
-                                imageService.findById(id);
-                                postImageService.create(new CreatePostImageRequestDTO(post.getId(), id));
-                            } catch(ImageNotFound e) {
-                            }
-                        });
-
+        PostResponseDTO postResponseDTO = postService.create(createPostRequestDTO, userId);
         return ResponseEntity.ok(postResponseDTO);
     }
 
@@ -98,21 +83,6 @@ public class PostController {
                 throw new IllegalArgumentException("태그는 영문자만 가능하고, 띄어쓰기를 포함할 수 없습니다: " + tag);
             }
         }
-    }
-
-    private List<UUID> extractImageIdsFromContent(String content) {
-        List<UUID> imageUUIDs = new ArrayList<>();
-
-        String uuidRegex = "\\{\\{([0-9a-fA-F-]+)\\|";
-        Pattern pattern = Pattern.compile(uuidRegex);
-        Matcher matcher = pattern.matcher(content);
-
-        while (matcher.find()) {
-                imageUUIDs.add(UUID.fromString(matcher.group(1)));
-        }
-
-        System.out.println(imageUUIDs.get(0));
-        return imageUUIDs;
     }
 
     private PostResponseDTO entityToDTO(Post post) {
