+++ src/main/java/com/sb02/blogpractice/service/PostService.java
@@ -1,6 +1,7 @@
 package com.sb02.blogpractice.service;
 
 import com.sb02.blogpractice.dto.CreatePostRequestDTO;
+import com.sb02.blogpractice.dto.PostResponseDTO;
 import com.sb02.blogpractice.dto.PostResponseListDTO;
 import com.sb02.blogpractice.dto.UpdatePostRequestDTO;
 import com.sb02.blogpractice.entity.Post;
@@ -10,7 +11,7 @@ import java.util.List;
 import java.util.UUID;
 
 public interface PostService {
-    Post create(CreatePostRequestDTO createRequestPostDTO, String userId);
+    PostResponseDTO create(CreatePostRequestDTO createRequestPostDTO, String userId);
     Post findById(UUID id);
     PostResponseListDTO findAll(int page, int size);
     Post update(UUID id, UpdatePostRequestDTO updateRequestPostDTO);
